Rename files to reflect class name
