+++ scopesim/effects/__init__.py
@@ -6,7 +6,7 @@ from .shifts import *
 
 from .obs_strategies import *
 from .spectral_trace_list import *
-from .spectral_efficiencies import *
+from .spectral_efficiency import *
 from .metis_lms_trace_list import *
 from .surface_list import *
 from .ter_curves import *
