+++ scopesim/effects/spectral_efficiency.py
