+++ scopesim/tests/tests_effects/test_SpectralEfficiency.py
